requirements.txt
Django==3.2.5
django-ckeditor==6.1.0
pre-commit==2.13.0
Pillow==8.3.2
six==1.16.0
social-auth-app-django==5.0.0
